requirements.txt
requests>=2.28.0
mcp
python-dotenv>=1.0.0